Separates activation function exceptions.
