+++ test/test_activation_functions.cpp
@@ -1,4 +1,5 @@
 #include "activation_functions.hpp"
+#include "exceptions/activation_functions.hpp"
 #include "utils/exceptions.hpp"
 #include <Eigen/Dense>
 #include <gtest/gtest.h>
